+++ application/io/exchange_coord_axes.py
@@ -1,5 +1,4 @@
-"""Exchange coordinate axes of two-dimensional three-point clustering
-statistics.
+"""Exchange coordinate axes of 2-d three-point clustering statistics.
 
 Examples
 --------
@@ -50,31 +49,39 @@ def configure():
     return config
 
 
-def transpose_filename(filepath):
-    """Transpose the filename of a two-dimensional three-point clustering
-    measurement file.
+def transpose_stat_str(pathlike, ret_multipole=False):
+    """Transpose the multipole indices in a 2-d three-point clustering
+    measurement path-like object.
 
     Parameters
     ----------
-    filepath : str
-        The file path of the input file.
+    pathlike : path-like
+        Path-like containing the multipole indices in the first of
+        multiple fields delimited by underscores.
+    ret_multipole : bool, optional
+        If `True` (default is `False`), return the original
+        multipole indices.
 
     Returns
     -------
-    filename_ex : :class:`pathlib.Path`
+    filenamelike_ex : :class:`pathlib.Path`
         The transposed filename.
+    stat_type : str
+        The statistic type, e.g. 'bk'.
+    multipole_ex : str
+        The transposed multipole indices.
+    multipole : str, optional
+        The original multipole indices.
 
     """
-    filepath = Path(filepath)
-    filename = filepath.stem
-    extension = filepath.suffix
+    pathlike = Path(pathlike)
 
-    filename_fields = filename.split('_')
-    stat_field = filename_fields[0]
-    nonstat_fields = filename_fields[1:]
+    fields = pathlike.stem.split('_')
+    stat_field, nonstat_fields = fields[0], fields[1:]
 
     stat_type = re.match(r'\D+', stat_field).group()
     stat_indices = re.findall(r'\d+', stat_field)
+    index_sep = ',' if ',' in stat_field else ''
 
     if len(stat_indices) == 1:
         stat_indices = list(stat_indices.pop())
@@ -84,23 +91,26 @@ def transpose_filename(filepath):
             "indices of length 3 before any underscores, "
             "e.g. 'bk000_' or 'bk0,0,0_'."
         )
+    if stat_indices[0] == stat_indices[1]:
+        raise ValueError("The first two multipole indices are the same.")
     stat_indices_ex = stat_indices[1:] + [stat_indices[0]]
 
-    if ',' in stat_field:
-        index_sep = ','
-    else:
-        index_sep = ''
-    stat_field_ex = stat_type + index_sep.join(stat_indices_ex)
+    multipole = index_sep.join(stat_indices)
+    multipole_ex = index_sep.join(stat_indices_ex)
 
-    filename_stem = '_'.join([stat_field_ex] + nonstat_fields)
-    filename_ex = Path(filename_stem).with_suffix(extension)
+    stat_field_ex = stat_type + multipole_ex
 
-    return filename_ex
+    pathlike_stem_ex = '_'.join([stat_field_ex] + nonstat_fields)
+    filenamelike_ex = Path(pathlike_stem_ex).with_suffix(pathlike.suffix)
+
+    if ret_multipole:
+        return filenamelike_ex, stat_type, multipole_ex, multipole
+    return filenamelike_ex, stat_type, multipole_ex
 
 
 def transpose_datatab(datatab):
-    """Transpose the data table of a two-dimensional three-point
-    clustering measurement file.
+    """Transpose the data table of a 2-d three-point clustering
+    measurement file.
 
     Parameters
     ----------
@@ -120,8 +130,8 @@ def transpose_datatab(datatab):
     coord_colidx_a = 0
     coord_colidx_b = 3
 
-    coord_cols_a = slice(coord_colidx_a, coord_colidx_a + coord_colspan - 1)
-    coord_cols_b = slice(coord_colidx_b, coord_colidx_b + coord_colspan - 1)
+    coord_cols_a = slice(coord_colidx_a, coord_colidx_a + coord_colspan)
+    coord_cols_b = slice(coord_colidx_b, coord_colidx_b + coord_colspan)
 
     datatab_ex[:, coord_cols_a] = datatab[:, coord_cols_b]
     datatab_ex[:, coord_cols_b] = datatab[:, coord_cols_a]
@@ -144,14 +154,27 @@ def main():
     cfg = configure()
 
     # Transpose the measurement filename.
-    output_filename = transpose_filename(cfg.input_path)
+    try:
+        output_filename, stat_type, multipole_ex, multipole = \
+            transpose_stat_str(cfg.input_path, ret_multipole=True)
+    except ValueError as err:
+        if "The first two multipole indices the same." in str(err):
+            print(
+                "No transposition is performed as the first two multipole "
+                "indices are the same."
+            )
+            return
+        # STYLE: Un-Pythonic 'else' clause after 'return' statement
+        # but retained for clarity.
+        else:
+            raise err
     if not cfg.output_dir:
         output_dir = Path(cfg.input_path).parent
     else:
         output_dir = Path(cfg.output_dir)
     if cfg.output_tag:
         if output_filename.suffix not in ('.dat', '.txt', '.csv', '.tsv'):
-            output_path = output_dir/(output_filename + cfg.output_tag)
+            output_path = output_dir/(str(output_filename) + cfg.output_tag)
         else:
             output_path = output_dir/(
                 output_filename.stem + cfg.output_tag + output_filename.suffix
@@ -166,7 +189,10 @@ def main():
         while line.startswith('#'):
             headerlines.append(line)
             line = measurement_file.readline()
-    header = ''.join(headerlines)
+    headerlines[-1] = headerlines[-1].replace(
+        stat_type + multipole, stat_type + multipole_ex
+    )
+    header = ''.join(headerlines).rstrip()
 
     # Transpose the data table.
     datatab = np.loadtxt(cfg.input_path)
@@ -177,7 +203,7 @@ def main():
         (['%.9e'] * 2 + ['%10d']) * 2 +
         ['% .9e'] * (datatab.shape[-1] - 6)
     )
-    np.savetxt(output_path, datatab_ex, fmt=fmt, header=header)
+    np.savetxt(output_path, datatab_ex, fmt=fmt, header=header, comments='')
     print(f"Transpose-exchanged measurement file saved to: {output_path}")