+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 - Fix FFTW-related memory leaks.
 
-- Fix asymmetric 2-dimensional three-point clustering statistics.
+- Fix asymmetric 2-d three-point clustering statistics.
 
 ### Features